fixed files not purging. new display setup, and added route tracking mode (initial)

+++ lib/storage/src/storage.h
@@ -8,6 +8,6 @@ void initStorage();
 void writeToLogFile(double lat, double lng, double speed);
 void writeToTimeLog(unsigned long lapTime, unsigned long sector1, unsigned long sector2, unsigned long sector3);
 void endSession();
-size_t storageUsage();
+double storageUsage();
 
 #endif
